linux-fslc: designate 6.1.y kernel as LTS
According to [1], Kernel version 6.1.y is designated as latest LTS and
there are newer Kernel versions available which shall be used as a base for
latest linux-fslc.

Rename linux-fslc_6.1 recipe to linux-fslc-lts_6.1 to designate it as LTS
and provide in a separate package.

Extend recipe description to reflect that this recipe provides LTS-based
Kernel.

Link: [1]: https://www.kernel.org/category/releases.html
Signed-off-by: Andrey Zhizhikin <andrey.z@gmail.com>