imx-boot: fix firmware name handling
In imx-boot's iMX9/soc.mak, the firmware name is set to a constant
value, i.e. mx93a0-ahab-container.img.

Instead, it should be set to SECO_FIRMWARE_NAME. This variable is
defined in use-imx-security-controller-firmware.bbclass, then it
is passed as a make parameter in imx-boot's do_compile() task.

As of now, there are revisions beyond A0 (namely A1), so the firmware
must be compiled accordingly. This patch implements such feature.

Signed-off-by: Joao Marcos Costa <joaomarcos.costa@bootlin.com>