imx-boot: Fix the recipe for when UBOOT_CONFIG is empty
The patchset to enable multiple u-boot targets (and as a consequence,
multimple imx-boot binaries) for the same machine is based only on
UBOOT_CONFIG variable.

For the case where a machine is not using UBOOT_MACHINE instead (and
UBOOT_CONFIG is empty) the previous code does not work, this is the
reason for this fix.

Signed-off-by: Daiane Angolini <daiane.angolini@foundries.io>