imx-gpu-viv: Normalize packaging
There are no behavioral changes here, just a cleanup to make packaging
lines consistent.

Note that SOLIBS and SOLIBSDEV are set the same as ".so", but the
former is used for main packages and the latter for -dev packages.

Also, drop ${libdir}/libGLESv1*${SOLIBS} from FILES:libgles1-imx-dev.
The rule has no effect since it is also included, correctly, in the
main package FILES:libgles1-imx.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>