imx-test: Use ROOT_HOME instead of hardcoded home directory
Hardcoding the root home directory could cause issues when the ROOT_HOME
is not set to '/home/root'. As example, the ROOT_HOME can be changed to
'/root', which makes imx-test install files in the wrong folder.

Therefore, replace '/home/root' with '$ROOT_HOME'.
Usually, ROOT_HOME is already set to '/home/root' [1].

[1] https://docs.yoctoproject.org/4.3.3/ref-manual/variables.html?highlight=variables#term-ROOT_HOME

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>