weston-init: Fix RDP support
The modification of the weston.ini command line necessary for RDP
support is missing. The problem is the weston.ini template contains
`@bindir@`, while the sed operation for the modification contains
`${bindir}`.

Fix the problem by moving the replacement of `@bindir@` to the top of
the function `do_install:append()`.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>