imx-boot: Fix 8M multi-config build problems
Building in the NXP layer meta-imx with the latest imx-boot recipe, the
dtb is not found during mkimage build:
```
| NOTE: building iMX8MP - TEE=tee.bin-stmm flash_evk_stmm_capsule
| dtc -@ -I dts -O dtb -o signature.dtbo signature.dts
| fdtoverlay -i imx8mp-evk.dtb -o imx8mp-evk.dtb signature.dtbo
| Couldn't open blob from 'imx8mp-evk.dtb': No such file or directory
```

The problem is the recipe does now copy the dtb from u-boot-imx with a
config-adorned name, e.g., imx8mp-evk.dtb-sd, but mkimage is still
expecting imx8mp-evk.dtb.

Add a link for compatibility.

Signed-off-by: Tom Hochstein <tom.hochstein@nxp.com>