imx-imx-boot-bootpart.wks.in: Set fixed size for /boot partition
Using --size in the imx-imx-boot-bootpart.wks.in to set size of /boot
partition does not result in a 256MiB partition as expected but in a
~332MiB one.  Indeed, --size ensures --extra-space (default 10MiB) is
added to actual data size and furthermore applies --overhead-factor
(default 1.3) to the wanted size. This is exactly the case here as 256
* 1.3 = 332.8.

Use --fixed-size instead to force the /boot partition to be exactly
256MiB just as it is indicated in the comments of this file.
Consequently, the `do_image_wic` task will also fail if data copied in
this partition are too large at creation.

Signed-off-by: Jonathan GUILLOT <jonathan@joggee.fr>
(cherry picked from commit 0ee7c736c7e07ce353e7f25ff514e9c459448103)