firmware-nxp-wifi: create symbolic links to firmware/mrvl folder
The upstream kernel Wi-Fi driver expects the sdiouart8997_combo_v4 and
sdiouartiw416_combo_v0 binaries to be located under firmware/mrvl [1],
different from the out-of-tree driver, which places them under
firmware/nxp. Create symbolic links to deploy these binaries in the
correct location for the upstream driver.

Due to this change, bump the recipe version to 1.1.

[1] https://github.com/torvalds/linux/blob/master/drivers/net/wireless/marvell/mwifiex/sdio.h#L31

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>