firmware-nxp-wifi: remove COMPATIBLE_MACHINE
The current COMPATIBLE_MACHINE is set to build every i.MX SoC with both
BSP variants. However the firmware-nxp-wifi deploys firmware related to
the NXP Wi-Fi Modules (e.g. Maya IW416), which can be used with other
SoCs as well, as example the Toradex Verdin-AM62 wich uses the TI SoC
and the NXP Wi-Fi Module.

With the current COMPATIBLE_MACHINE it is not possible to build this
recipe and deploy the NXP firmware with different SoCs, therefore remove
it.

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>