alsa-lib: rework patches to apply on 1.2.13 and 1.2.14
OE-core updated alsa-lib.
commit 8518e57a117a ("alsa-lib: upgrade 1.2.13 -> 1.2.14")

The patches in their current form either apply to one or the
other version without fuzz.
Move the additions to the cfg_files variable list to the top
rather than to the bottom of the list to have the patches apply
cleanly to the older and newer alsa-lib version.

Fixes: 32226986810a ("alsa-lib: refresh patches") on walnascar
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>