imx-base.inc: Update i.MX95 default Rev to B0
This is necessary to align with NXP BSP LF6.6.52-2.2.1.

In imx-oei, the default DDR timings have been updated to B0, and anyone
building for A0/A1 is expected to pass the argument DDR_CONFIG with the
A1 timings to imx-oei.

Therefore, a machine conf for A0/A1 support that also sets the desired
timings is always necessary, and we have to drop A0/A1 from
imx-base.inc. It is not enough to override IMX_SOC_REV to support A0/A1.

Signed-off-by: Ernest Van Hoecke <ernest.vanhoecke@toradex.com>